Move IPv4 unit test cases from IP folders to IPv4 folders.
