+++ test/unit-test/FreeRTOS_IP_DiffConfig/ut.cmake
@@ -17,6 +17,7 @@ list(APPEND mock_list
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP_Timers.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP_Utils.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IPv4_Private.h"
+            "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IPv4.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_Routing.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_ARP.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_ICMP.h"
@@ -58,7 +59,6 @@ set(real_source_files "")
 list(APPEND real_source_files
             ${project_name}/${project_name}_stubs.c
             ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_IP.c
-            ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_IPv4.c
 	)
 
 set(real_include_directories "")
