+++ test/unit-test/FreeRTOS_IP_DiffConfig2/ut.cmake
@@ -16,6 +16,7 @@ list(APPEND mock_list
                 "${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include/event_groups.h"
                 "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_ND.h"
                 "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_Routing.h"
+                "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IPv4.h"
                 "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IPv4_Private.h"
                 "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP_Timers.h"
                 "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP_Utils.h"
@@ -59,7 +60,6 @@ set(real_source_files "")
 list(APPEND real_source_files
             ${project_name}/${project_name}_stubs.c
             ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_IP.c
-            ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_IPv4.c
 	)
 
 set(real_include_directories "")
