+++ test/unit-test/FreeRTOS_IP_DiffConfig3/FreeRTOS_IP_DiffConfig3_utest.c
@@ -59,6 +59,7 @@
 #include "mock_FreeRTOS_Stream_Buffer.h"
 #include "mock_FreeRTOS_TCP_WIN.h"
 #include "mock_FreeRTOS_UDP_IP.h"
+#include "mock_FreeRTOS_IPv4.h"
 
 #include "FreeRTOS_IP.h"
 
@@ -114,453 +115,6 @@ static size_t StubuxStreamBufferGetPtr_ReturnCorrectVals( StreamBuffer_t * pxBuf
     *ppucData = &ReleaseTCPPayloadBuffer[ 0 ];
 
     return ReleaseTCPPayloadBufferxByteCount;
-}
-
-void test_prvAllowIPPacketIPv4_BufferLengthLess( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x46;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxIPHeader->usLength = FreeRTOS_htons( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( TCPHeader_t ) );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    for( uint32_t i = 0; i < sizeof( IPPacket_t ); i++ )
-    {
-        pxNetworkBuffer->xDataLength = i;
-        eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-        TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-    }
-}
-
-void test_prvAllowIPPacketIPv4_HeaderLengthLess( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxNetworkBuffer->xDataLength = sizeof( TCPPacket_t );
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45 - 2;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxIPHeader->usLength = FreeRTOS_htons( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( TCPHeader_t ) );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_HeaderLengthMore( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxNetworkBuffer->xDataLength = sizeof( TCPPacket_t );
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F + 2;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxIPHeader->usLength = FreeRTOS_htons( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( TCPHeader_t ) );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_HeaderLengthMoreThanTotalLength( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 );
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxIPHeader->usLength = FreeRTOS_htons( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( TCPHeader_t ) );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_IPPacketLengthMoreThanTotalLength( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxIPHeader->usLength = FreeRTOS_htons( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( TCPHeader_t ) );
-    pxNetworkBuffer->xDataLength = FreeRTOS_ntohs( pxIPHeader->usLength ) - 1;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_UDP_IncorrectPacketLen( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_UDP;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_UDP_HEADER - 1;
-    pxIPHeader->usLength = FreeRTOS_htons( pxNetworkBuffer->xDataLength - ipSIZE_OF_ETH_HEADER );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_TCP_IncorrectPacketLen( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_TCP;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_TCP_HEADER - 1;
-    pxIPHeader->usLength = FreeRTOS_htons( pxNetworkBuffer->xDataLength - ipSIZE_OF_ETH_HEADER );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_ICMP_IncorrectPacketLen( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_ICMP;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_ICMPv4_HEADER - 1;
-    pxIPHeader->usLength = FreeRTOS_htons( pxNetworkBuffer->xDataLength - ipSIZE_OF_ETH_HEADER );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_IGMP_IncorrectPacketLen( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_IGMP;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_ICMPv4_HEADER - 1;
-    pxIPHeader->usLength = FreeRTOS_htons( pxNetworkBuffer->xDataLength - ipSIZE_OF_ETH_HEADER );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_NoProt( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = 0;
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_ICMPv4_HEADER - 1;
-    pxIPHeader->usLength = FreeRTOS_htons( pxNetworkBuffer->xDataLength - ipSIZE_OF_ETH_HEADER );
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_UDP_LengthLess( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_UDP;
-    pxIPHeader->usLength = FreeRTOS_ntohs( ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + sizeof( pxProtPack->xUDPPacket.xUDPHeader ) - 1 );
-    pxNetworkBuffer->xDataLength = ( ( pxIPHeader->ucVersionHeaderLength & 0x0F ) << 2 ) + ipSIZE_OF_ETH_HEADER + ipSIZE_OF_UDP_HEADER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_UDP_LengthMore( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = ipSIZE_OF_IPv4_HEADER;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-    ProtocolPacket_t * pxProtPack;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    /* Set correct length. */
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x4F;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-    /* Correct protocol. */
-    pxIPHeader->ucProtocol = ipPROTOCOL_UDP;
-    pxIPHeader->usLength = FreeRTOS_ntohs( ipconfigNETWORK_MTU + 1 );
-    pxNetworkBuffer->xDataLength = ipconfigNETWORK_MTU + 20;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
 }
 
 void test_prvProcessIPEventsAndTimers_eDHCPEvent_DHCPv4( void )
