+++ test/unit-test/FreeRTOS_IPv4/FreeRTOS_IPv4_utest.c
@@ -49,7 +49,6 @@
 
 #include "FreeRTOS_IPv4.h"
 
-#include "FreeRTOS_IPv4_stubs.c"
 #include "catch_assert.h"
